backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
python-dotenv
pandas
openai
langdetect
pdfplumber
opencv-python-headless
Pillow
requests
transformers
peft
accelerate
torch
sentencepiece
protobuf
bitsandbytes

# Note: This project uses a remote PaddleOCR API by default (configured via
# the PADDLE_OCR_URL environment variable in backend/.env). Local `paddleocr`
# installation is optional — see backend/PADDLE_INSTALL.md if you want a
# self-hosted/local Paddle setup.
pdf2image


